Refactored VectoredIO interface for VectoredIOImpl and added flag.

+++ gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageFileSystemImpl.java
@@ -38,6 +38,7 @@ import com.google.cloud.hadoop.util.LazyExecutorService;
 import com.google.cloud.hadoop.util.ThreadTrace;
 import com.google.cloud.hadoop.util.TraceOperation;
 import com.google.common.annotations.VisibleForTesting;
+import com.google.common.base.Strings;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterables;
@@ -51,8 +52,11 @@ import java.nio.channels.WritableByteChannel;
 import java.nio.file.DirectoryNotEmptyException;
 import java.nio.file.FileAlreadyExistsException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.Comparator;
+import java.util.HashMap;
+import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -68,6 +72,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 /** Provides FS semantics over GCS based on Objects API */
@@ -327,6 +332,9 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
               () -> getFileInfoInternal(parentId, /* inferImplicitDirectories= */ false));
     }
 
+    boolean isHnBucket =
+        (this.options.getCloudStorageOptions().isHnBucketRenameEnabled() && gcs.isHnBucket(path));
+    List<FolderInfo> listOfFolders = new LinkedList<>();
     List<FileInfo> itemsToDelete;
     // Delete sub-items if it is a directory.
     if (fileInfo.isDirectory()) {
@@ -338,6 +346,32 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
               : listFileInfoForPrefixPage(
                       fileInfo.getPath(), DELETE_RENAME_LIST_OPTIONS, /* pageToken= */ null)
                   .getItems();
+
+      /*TODO : making listing of folder and object resources in parallel*/
+      if (isHnBucket) {
+        /**
+         * Get list of folders if the bucket is HN enabled bucket For recursive delete, get all
+         * folder resources. For non-recursive delete, delete the folder directly if it is a
+         * directory and do not do anything if the path points to a bucket.
+         */
+        String bucketName = getBucketName(path);
+        String folderName = getFolderName(path);
+        listOfFolders =
+            recursive
+                ? listFoldersInfoForPrefixPage(
+                        fileInfo.getPath(), ListFolderOptions.DEFAULT, /* pageToken= */ null)
+                    .getItems()
+                // will not delete for a bucket
+                : (folderName.equals("")
+                    ? new LinkedList<>()
+                    : Arrays.asList(
+                        new FolderInfo(FolderInfo.createFolderInfoObject(bucketName, folderName))));
+
+        logger.atFiner().log(
+            "Encountered HN enabled bucket with %s number of folder in path : %s",
+            listOfFolders.size(), path);
+      }
+
       if (!itemsToDelete.isEmpty() && !recursive) {
         GoogleCloudStorageEventBus.postOnException();
         throw new DirectoryNotEmptyException("Cannot delete a non-empty directory.");
@@ -349,21 +383,92 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
     List<FileInfo> bucketsToDelete = new ArrayList<>();
     (fileInfo.getItemInfo().isBucket() ? bucketsToDelete : itemsToDelete).add(fileInfo);
 
-    deleteInternal(itemsToDelete, bucketsToDelete);
+    deleteInternalWithFolders(itemsToDelete, listOfFolders, bucketsToDelete);
 
     repairImplicitDirectory(parentInfoFuture);
   }
 
+  /**
+   * Return the bucket name if exists
+   *
+   * @param path
+   * @return bucket name
+   */
+  private String getBucketName(@Nonnull URI path) {
+    checkState(
+        !Strings.isNullOrEmpty(path.getAuthority()), "Bucket name cannot be null : %s", path);
+    return path.getAuthority();
+  }
+
+  /**
+   * Returns the folder name if exists else return empty string.
+   *
+   * @param path
+   * @return folder path
+   */
+  private String getFolderName(@Nonnull URI path) {
+    checkState(
+        path.getPath().startsWith(PATH_DELIMITER), "Invalid folder name: %s", path.getPath());
+    return path.getPath().substring(1);
+  }
+
+  /**
+   * Returns the list of folder resources in the prefix. It lists all the folder resources
+   *
+   * @param prefix the prefix to use to list all matching folder resources.
+   * @param pageToken the page token to list
+   * @param listFolderOptions the page token to list
+   */
+  public ListPage<FolderInfo> listFoldersInfoForPrefixPage(
+      URI prefix, ListFolderOptions listFolderOptions, String pageToken) throws IOException {
+    logger.atFiner().log(
+        "listFoldersInfoForPrefixPage(prefix: %s, pageToken:%s)", prefix, pageToken);
+    StorageResourceId prefixId = getPrefixId(prefix);
+    return gcs.listFolderInfoForPrefixPage(
+        prefixId.getBucketName(), prefixId.getObjectName(), listFolderOptions, pageToken);
+  }
+
+  /**
+   * Deletes the given folder resources
+   *
+   * @param listOfFolders to delete
+   * @throws IOException
+   */
+  private void deleteFolders(@Nonnull List<FolderInfo> listOfFolders) throws IOException {
+    if (listOfFolders.isEmpty()) return;
+    logger.atFiner().log(
+        "deleteFolder(listOfFolders: %s, size:%s)", listOfFolders, listOfFolders.size());
+    gcs.deleteFolders(listOfFolders);
+  }
+
+  /** Deletes all objects in the given path list followed by all bucket items. */
+
   /** Deletes all items in the given path list followed by all bucket items. */
   private void deleteInternal(List<FileInfo> itemsToDelete, List<FileInfo> bucketsToDelete)
       throws IOException {
+
+    deleteObjects(itemsToDelete);
+    deleteBucket(bucketsToDelete);
+  }
+
+  /** Deleted all objects, folders and buckets in the order mentioned */
+  private void deleteInternalWithFolders(
+      List<FileInfo> itemsToDelete, List<FolderInfo> listOfFolders, List<FileInfo> bucketsToDelete)
+      throws IOException {
+    deleteObjects(itemsToDelete);
+    deleteFolders(listOfFolders);
+    deleteBucket(bucketsToDelete);
+  }
+
+  /** Helper function to delete objects */
+  private void deleteObjects(List<FileInfo> itemsToDelete) throws IOException {
     // TODO(user): We might need to separate out children into separate batches from parents to
     // avoid deleting a parent before somehow failing to delete a child.
 
     // Delete children before their parents.
     //
     // Note: we modify the input list, which is ok for current usage.
-    // We should make a copy in case that changes in the future.
+    // We should make a copy in case that changes in future.
     itemsToDelete.sort(FILE_INFO_PATH_COMPARATOR.reversed());
 
     if (!itemsToDelete.isEmpty()) {
@@ -380,7 +485,10 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
       }
       gcs.deleteObjects(objectsToDelete);
     }
+  }
 
+  /** Helper function to delete buckets */
+  private void deleteBucket(List<FileInfo> bucketsToDelete) throws IOException {
     if (!bucketsToDelete.isEmpty()) {
       List<String> bucketNames = new ArrayList<>(bucketsToDelete.size());
       for (FileInfo bucketInfo : bucketsToDelete) {
@@ -507,14 +615,25 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
           StorageResourceId.fromUriPath(
               dst, /* allowEmptyObjectName= */ true, /* generationId= */ 0L);
 
-      gcs.copy(ImmutableMap.of(srcResourceId, dstResourceId));
+      if (this.options.getCloudStorageOptions().isMoveOperationEnabled()
+          && srcResourceId.getBucketName().equals(dstResourceId.getBucketName())) {
+        gcs.move(
+            ImmutableMap.of(
+                new StorageResourceId(
+                    srcInfo.getItemInfo().getBucketName(),
+                    srcInfo.getItemInfo().getObjectName(),
+                    srcInfo.getItemInfo().getContentGeneration()),
+                dstResourceId));
+      } else {
+        gcs.copy(ImmutableMap.of(srcResourceId, dstResourceId));
 
-      gcs.deleteObjects(
-          ImmutableList.of(
-              new StorageResourceId(
-                  srcInfo.getItemInfo().getBucketName(),
-                  srcInfo.getItemInfo().getObjectName(),
-                  srcInfo.getItemInfo().getContentGeneration())));
+        gcs.deleteObjects(
+            ImmutableList.of(
+                new StorageResourceId(
+                    srcInfo.getItemInfo().getBucketName(),
+                    srcInfo.getItemInfo().getObjectName(),
+                    srcInfo.getItemInfo().getContentGeneration())));
+      }
     }
 
     repairImplicitDirectory(srcParentInfoFuture);
@@ -657,6 +776,29 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
       }
     }
 
+    StorageResourceId srcResourceId =
+        StorageResourceId.fromUriPath(src, /* allowEmptyObjectName= */ true);
+    StorageResourceId dstResourceId =
+        StorageResourceId.fromUriPath(
+            dst, /* allowEmptyObjectName= */ true, /* generationId= */ 0L);
+    if (this.options.getCloudStorageOptions().isMoveOperationEnabled()
+        && srcResourceId.getBucketName().equals(dstResourceId.getBucketName())) {
+
+      // First, move all items except marker items
+      moveInternal(srcToDstItemNames);
+      // Finally, move marker items (if any) to mark rename operation success
+      moveInternal(srcToDstMarkerItemNames);
+
+      if (srcInfo.getItemInfo().isBucket()) {
+        deleteBucket(Collections.singletonList(srcInfo));
+      } else {
+        // If src is a directory then srcItemInfos does not contain its own name,
+        // we delete item separately in the list.
+        deleteObjects(Collections.singletonList(srcInfo));
+      }
+      return;
+    }
+
     // First, copy all items except marker items
     copyInternal(srcToDstItemNames);
     // Finally, copy marker items (if any) to mark rename operation success
@@ -706,6 +848,27 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
 
     // Perform copy.
     gcs.copy(srcBucketName, srcObjectNames, dstBucketName, dstObjectNames);
+  }
+
+  /** Moves items in given map that maps source items to destination items. */
+  private void moveInternal(Map<FileInfo, URI> srcToDstItemNames) throws IOException {
+    if (srcToDstItemNames.isEmpty()) {
+      return;
+    }
+
+    Map<StorageResourceId, StorageResourceId> sourceToDestinationObjectsMap = new HashMap<>();
+
+    // Prepare list of items to move.
+    for (Map.Entry<FileInfo, URI> srcToDstItemName : srcToDstItemNames.entrySet()) {
+      StorageResourceId srcResourceId = srcToDstItemName.getKey().getItemInfo().getResourceId();
+
+      StorageResourceId dstResourceId =
+          StorageResourceId.fromUriPath(srcToDstItemName.getValue(), true);
+      sourceToDestinationObjectsMap.put(srcResourceId, dstResourceId);
+    }
+
+    // Perform move.
+    gcs.move(sourceToDestinationObjectsMap);
   }
 
   /**
@@ -866,6 +1029,17 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
   }
 
   @Override
+  public FileInfo getFileInfoWithHint(URI path, PathTypeHint pathTypeHint) throws IOException {
+    checkArgument(path != null, "path must not be null");
+    StorageResourceId resourceId = StorageResourceId.fromUriPath(path, true);
+    FileInfo fileInfo =
+        FileInfo.fromItemInfo(
+            getFileInfoInternal(resourceId, /* inferImplicitDirectories= */ true, pathTypeHint));
+    logger.atFiner().log("getFileInfo(path: %s): %s", path, fileInfo);
+    return fileInfo;
+  }
+
+  @Override
   public FileInfo getFileInfoObject(URI path) throws IOException {
     checkArgument(path != null, "path must not be null");
     StorageResourceId resourceId = StorageResourceId.fromUriPath(path, true);
@@ -879,18 +1053,26 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
     return fileInfo;
   }
 
+  private GoogleCloudStorageItemInfo getFileInfoInternal(
+      StorageResourceId resourceId, boolean inferImplicitDirectories) throws IOException {
+    return getFileInfoInternal(resourceId, inferImplicitDirectories, PathTypeHint.NONE);
+  }
+
   /**
    * @see #getFileInfo(URI)
    */
   private GoogleCloudStorageItemInfo getFileInfoInternal(
-      StorageResourceId resourceId, boolean inferImplicitDirectories) throws IOException {
+      StorageResourceId resourceId, boolean inferImplicitDirectories, PathTypeHint pathTypeHint)
+      throws IOException {
     if (resourceId.isRoot() || resourceId.isBucket()) {
       return gcs.getItemInfo(resourceId);
     }
-    StorageResourceId dirId = resourceId.toDirectoryId();
 
+    StorageResourceId dirId = resourceId.toDirectoryId();
     Future<List<GoogleCloudStorageItemInfo>> listDirFuture =
-        (options.isStatusParallelEnabled() ? cachedExecutor : lazyExecutor)
+        (options.isStatusParallelEnabled() && pathTypeHint != PathTypeHint.FILE
+                ? cachedExecutor
+                : lazyExecutor)
             .submit(
                 () ->
                     inferImplicitDirectories
@@ -1095,5 +1277,10 @@ public class GoogleCloudStorageFileSystemImpl implements GoogleCloudStorageFileS
   @Override
   public GoogleCloudStorage getGcs() {
     return gcs;
+  }
+
+  public enum PathTypeHint {
+    FILE,
+    NONE
   }
 }
